pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sales-coach"
version = "0.1.0"
description = "Local AI-powered sales coaching system"
authors = [
    {name = "Sales Coach Team", email = "team@salescoach.local"},
]
dependencies = [
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "librosa>=0.10.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.30.0",
    "whisper-cpp-python>=0.1.0",
    "pyannote.audio>=3.1.0",
    "silero-vad>=4.0.0",
    "llama-cpp-python>=0.2.0",
    "accelerate>=0.20.0",
    "psutil>=5.9.0",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
sales-coach = "sales_coach.cli:main"

[tool.setuptools.packages.find]
include = ["sales_coach*"]
exclude = ["bundles*", "config*", "logs*", "models_cache*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]